[Change] separate system controllers from routes
